refactor: 替换所有组件中的 Loader 导入为 LoadingPlaceholder，移除旧的 loader 组件
