+++ frontend/src/components/LoadingPlaceholder.tsx
