+++ frontend/src/components/d3/heatmap.tsx
@@ -7,7 +7,7 @@ import moment from "moment";
 import React from "react";
 import toast from "react-hot-toast";
 import { useTranslation } from "react-i18next";
-import LoadingPlaceholder from "../loader";
+import LoadingPlaceholder from "../LoadingPlaceholder";
 
 const Heatmap: React.FC = () => {
   const { t } = useTranslation("translation", {
