+++ frontend/src/components/tagMap.tsx
@@ -5,7 +5,7 @@ import { ArrowDownLeft } from "lucide-react";
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
 import { Link } from "react-router-dom";
-import LoadingPlaceholder from "./loader";
+import LoadingPlaceholder from "./LoadingPlaceholder";
 
 export default function TagMap() {
   const { t } = useTranslation("translation", {
